lttng add-trigger: remove `--userspace` and `--kernel` domain options
I believe the project is trying to steer away from those dedicated
tracing domain options, just like we did for the `event-rule-matches`
condition which includes a domain name as part of its `--type` option,
for example `--type=user:tracepoint` and `--type=log4j2:logging`.

Therefore remove the dedicated `--userspace` and `--kernel` options
which are not part of a release at this point.

Remove now-irrelevant tests in `test_add_trigger.py`.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I9896e7c3db2c1e99d815254ab187271a33d28f69
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>