Fix: Do not allow "data consumption paused" testpoint to skip blocking pipe reads
Move the consumerd data concumption paused testpoint _after_ handling
of the consumer_data_pipe. This pipe lets the consumer data thread know
about streams which are added and removed. It is a blocking pipe, so
failure to drain it may cause the sender to block.

The write-side of this pipe is within:

LTTNG_CONSUMER_GET_CHANNEL:
  send_streams_to_thread
    send_stream_to_thread

so it can effectively block the handling of a get channel command from
sessiond.

Note that this is only expected to happen if the pipe buffer is full.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I6ab6d58c665744779d0bd177925f3011adbace6e
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>