extras/zsh-completion/_lttng: remove __lttng_arguments()
While debugging this completion file, I found that, for some reason, if
I use another version of __lttng_arguments() with `-C`, like this:

    __lttng_arguments_c() {
      _arguments -s -W -C : "$@"
    }

then when calling it instead of exactly the same command line, I don't
see the usual updates to the `words` array when handling a state. For
example, with `lttng add-trigger <tab>`, `words` remains [`lttng`,
`add-trigger`] when calling __lttng_arguments_c() whereas with a direct
_arguments() call it's [`add-trigger`] (what I would expect).

I'm not sure to understand why, but because _arguments() seems to be
that capricious, I'm deciding to not take any risk and always use it
directly, even without `-C`.

This is not changing anything completion-wise, but reducing the risk of
misusing the completion system in the future.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I9758dea9098026a78d96dbaeed11afef300fffa7
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>