tests: util event generator: Start generator in suspended state
In run once mode, the test_app_function is only called once on util
event generator startup, so the execution of the test_app_function may
not happen on first resume, when expected.

Initialize the generator in suspended state, so the first transition to
resumed state trigger a test_app_function run.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: If1680b409f8cda9cbd24ddbcde553c43dead2beb
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>