Fix: test utils: clear sessiond pids array on stop_lttng_sessiond
Tests that start then stop the sessiond multiple times show errors when
trying to kill (again) the same pids that were already killed. This is a
real issue because this failing kill attempt may be killing unrelated
tasks.

This is also an issue for sigstop_lttng_sessiond_opt which also uses
that same list: it may try to sent "SIGSTOP" to unrelated PIDs by chance
and hang the test.

Fix this by clearing the pids array when stopping the sessiond.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I572cf024bd058d00f795f90135e27f926966f0b0
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>