build: Make invocation of clang2py more robust
In coverity builds, the build failed with the following error:

    02:14:14
    /home/jenkins/workspace/****_master_coverity/src/****/tests/vendor/ctypeslib2/ctypeslib/clang2py.py
    -i --clang-args="-I../../include -I../../src" -l
    /home/jenkins/workspace/****_master_coverity/src/****/src/lib/lttng-ctl/.libs/liblttng-ctl.so
    ../../include/lttng/lttng.h > liblttngctl/lttngctl/lttng.py
    02:14:14 Traceback (most recent call last):
    02:14:22   File "/home/jenkins/workspace/****_master_coverity/src/****/tests/vendor/ctypeslib2/ctypeslib/clang2py.py", line 18, in <module>
    02:14:22     import ctypeslib
    02:14:22 ModuleNotFoundError: No module named 'ctypeslib'
    make[3]: *** [Makefile:2331: liblttngctl/lttngctl/lttng.py] Error 1

I believe this occurs because cov-build is intercepting the
command. In our coverity jobs, we don't specify a specific
configuration[1] which I think means that all the configurations are
active by default.

Certainly in the build's workspace I can see the configuration for the
coverity python interceptor which matches `^.*\.p(y|y3)$`. As the
command is intercepted by cov-build, it's then run with by coverity but
with a slightly different environment.

The same issue can be seen when going to `tests/unit` and running
`make` from there - the command isn't specific enough.

The solution is to provide the exact command that should be run
including setting the environment variable appropriately.

[1]: https://github.com/lttng/lttng-ci/blob/dfd85485e29e09f7fef36b99874a493beeb8ab62/scripts/common/coverity.sh#L154

Change-Id: I382aec8f65cbbd80ea0bc956f76b0931b1800f17
Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>