unix socket: allow silencing on EPIPE
lttcomm_send_unix_sock is used to send messages through sockets
that may have been legitimately closed by the peer. For instance,
the liblttng-ctl protocol has a number of async commands that
clients may opt not to wait for.

In those cases, if the client process has exited, the session
daemon reports errors which spam the logs.

Change-Id: I1cc04d3e2aa34ba5b08586795ce55085c0d75f8e
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>