Fix: ust-consumer: missing subbuffer on explicit reclamation
When issuing a reclamation with an age limit, the "current" subbuffer
may be skipped if it was just flushed. This is because flushing
updates the subbuffer's timestamp, which causes it to fail the age
check.

To address this in the case when buffers must be consumed before
reclamation, we mark a subbuffer for deferred reclamation if it was
flushed during the reclamation process. This deferred reclamation is
then handled in the usual reclamation path when the consumer tries to
consume the current subbuffer.

However, it doesn't address the case where consumption is not required
(i.e., snapshot sessions).

Change-Id: I72b544eb1e7ecbb37c564fad2646045f19d49706
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>