doc: lttng-reclaim-memory(1): fix synopsis channel selection syntax
The synopsis showed `([--all] | CHANNEL...)` which incorrectly
suggested that `--all` is optional within the choice group.

Remove the brackets around `--all` so the synopsis correctly shows
`(--all | CHANNEL...)`, indicating that either `--all` or one or more
channel names must be specified.

Specifying nothing is reserved for a future "current channel"
in `.lttngrc`.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: Id7e222f93be8ccc29c4a67125e5b495b0b2286a4
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>